splits helpers for easier test mocking, adds unit tests for getIconFromTokenLists
