+++ extension/src/popup/helpers/assetList.ts
@@ -5,7 +5,7 @@ import {
 } from "@shared/constants/soroban/asset-list";
 import { getNativeContractDetails } from "popup/helpers/searchAsset";
 
-import { schemaValidatedAssetList } from "@shared/api/helpers/getIconFromTokenLists";
+import { schemaValidatedAssetList } from "@shared/api/helpers/token-list";
 import { ManageAssetCurrency } from "popup/components/manageAssets/ManageAssetRows";
 import { getAssetLists } from "./searchAsset";
